Update TunnelManager.java
fixed no internet connection on connect vpn

I've been debugging all day by trial and error and finally found the bug. Remember I told him to comment on this because it didn't work and I thought it was the app filter https://github.com/Andley302/SecondVPN/issues/1#issuecomment-2568625376

Well I wasn't 100% sure since I commented all the code but I took time and set myself the task of finding the bug. I recently told him that there was an error with address in use. Well the problem itself is related to this VPN function

//allowAllAFFamilies(builder);

Now I'm closing the case. I'll send you a commit with everything I modified and updated.
I changed the way that it excludes apps.
The current version of git slows down when scrolling quickly or scrolling back. The new implementation makes it more fluid and friendly.